## Why is this change being made?
This change will enable accelerator firmware download during MSCP warm reset sequence (Quiesce flow)

## What is changed?
Updated logic to enable

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (Test using HSP reset with the following command: "reset 3 3 0x9C 6 1")
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [x] Silicon
- [ ] Others, explain:

## Notes
![image.png](https://azurecsi.visualstudio.com/2610da6a-0168-48f8-b34a-f1c882cb6660/_apis/git/repositories/8fc5d0fc-7430-45b1-8963-5d075fba03da/pullRequests/140362/attachments/image.png)
## Risks

fix: Update accelerator quiesce flow to allow for accel code download.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a bug fix that updates the accelerator quiesce flow to properly support accel code download during warm reset.

#### PR Summary
The changes refine boot type handling and reset sequencing to enable correct operation of the quiesce flow, addressing back-to-back developer warm reset tests.
- `src/services/ap_core/ap_core.c`: Replaced the conditional check with a switch-case structure to better differentiate between boot types and handle the quiesce flow.
- `src/services/startup_shutdown/startup_shutdown_core_comm.c`: Removed the direct call to suspend the accelerator core, adding detailed TODO comments to outline new reset and interrupt masking steps.
- `src/libs/accelerator_ip/accelerator_ip_warm_reset.c`: Refactored functions by introducing `accel_emcpu_reset_sequence` to standardize the warm reset process for accelerators.
- `src/libs/accelerator_ip/accelerator_ip_init.c`: Added logic to reset the accelerator in SCP warm boot scenarios to ensure proper quiescence.
- `src/libs/accelerator_ip/inc/accelerator_ip.h` & `CMakeLists.txt`: Introduced a new API prototype and dependency to support the revised quiesce flow.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #2852927, #2853146

+++ src/services/startup_shutdown/startup_shutdown_thread.c
@@ -72,7 +72,7 @@ int sos_queue_find_phase(ssi_startup_stage_t phase)
 void sos_accel_quiesce_complete(uint32_t accel_id, uint32_t event_flag)
 {
     FPFW_UNUSED(accel_id)
-    SOS_LOG_TRACE("Accel:%d Quiesce Complete, event flag: %x\n", accel_id, event_flag);
+    SOS_LOG_TRACE("Accel:%d Quiesce Complete, event flag: %x\n", (int)accel_id, (int)event_flag);
     int status = tx_event_flags_set(&s_sos_thread_ctx.ssi_flags, event_flag, TX_OR);
     FPFW_RUNTIME_ASSERT(status == TX_SUCCESS);
 }
